exception.h: fix export header include to work in namespace-prefixed include
POSIX asks C compilers (and C++ compilers follow) for that "" includes
"shall be searched for first in the directory of the file with the
#include line", overruling the "The named source file is
searched for in an implementation-defined manner." of the C standard.

So to find the export header in the same dir in the installation setup,
while the class header itself is expected to be included with the
namespace-prefixed path via <threadweaver/exception.h> or
<Threadweaver/Exception>, one e.g. has to ue "", as already done by
the other headers. This works also in the build dir setup, as the
generated export header in the build dir will be found after checking
the same dir when searching then in the defined build include directories.